Base functionality of app with model data, tabs, navigation, and routing (still TODO: full screen login cover)
- Two tabs
    - Patients
    - Doctors
    - Each tab starts on the respective dashboard page

For now, dashboard page shows a button to navigate to the Patient or Doctor List.
When going into a Patient or Doctor List, you can click the entry and see the entry details.

Doctor have a Department property, which is used to section/sort the doctors in the Doctor List.

Adding a Patient/Doctor can be done using the toolbar button in the Patient/Doctor list;
adding a Patient/Doctor is done using a `.sheet`.

Still TODO:
- Add a Login Screen. This should be a fullscreen cover over the Tabs.
- The login screen route will be its own path, independent of PatientRoute, or DoctorRoute.
    - Username
    - Password
    - Confirm button
        - Leads to Connecting Screen
            - If credentials are correct, pop the fullscreen cover and reveal the tabs.
            - If incorrect, notify the user and provide a button to navigate back

After that, this project can serve as a base template for future apps like it
(login screen, then tab-based navigation).

Add more bells and whistles to make the app more interesting.
The idea here is to learn how to properly manage navigation in SwiftUI, particularly with TabView.