Remove code repetition of NavigationStacks -- double down on generics
Since all code for the `---NavigationStack` types was algorithmically the same,
why not genericize it?

The key was ensuring that types conforming to `Route` would be able to return the root view for the navigation route.
A static property would do.

Now we can remove all of the repetitious `--NavigationStack` types and have one `NavigationStackView` with a `RouteType: Route` type parameter.